Add files via upload
## Dataset

The dataset used in this project is a movie dataset (`mymovie.csv`) containing detailed information about movies and TV shows.  

### Features include:
- **Release_Date**: The date when the movie or TV show was released.
- **Title**: The title of the movie or show.
- **Popularity**: Popularity score of the movie or show.
- **Vote_Count**: Total number of votes received.
- **Vote_Average**: Average rating based on votes.
- **Genre**: Genre(s) assigned to the movie or show.

The dataset was uploaded manually for the purpose of this Netflix Data Analysis project.  
It will be used for exploratory data analysis, visualizations, and generating key insights about movie trends, genres, and popularity.

(*Note: This dataset is for educational and non-commercial research use only.*)